Use size check to quickly reject unequal files
In multiple-mode, it can be much faster to first verify that the files
cannot be equal by checking if their sizes don't match.  In that case,
it makes no sense to compute a hash for potentially large files.

Also, by checking size first, we ensure that if the hashes match, they
must also agree on file size as well (a much stronger assurance for
equivalency of the crypto-hashed files).