Ensure setup of MaxSize before hash computation
The hash computations use MaxSize as a to limit their work, such as if
given a device file with unbounded I/O (ie. /dev/zero).  The default
value of zero should be set before getHash attempts to compute a hash,
so that the hash input is not truncated to zero.

Rather than special handling for character devices, it may be preferable
for such cases to use CompareReader().  For example, what should the
expected result be for comparing /dev/zero to a 1K file file of all
zeros?  If MaxSize is not set, or set to > 1K, probably they should not
match.  And if MaxSize is set to zero, probably they should match.
However, Stat() will return size 0 for /dev/zero (and other character
device files), so arguably CompareFiles() is not well defined on
character device files (and that's just for Unix).