Ignore Opt.MaxSize if LimitedReader is being used
Allow io.LimitedReader to override/ignore Opt.MaxSize in
CompareReaders().  One LimitedReader is sufficient to provide a bound on
the amount of IO performed.

Passes the file size onto both getHash() and compareReader() to allow
detection of files that grow during the comparison, and also lets the
maxSize values to change as needed without actually modifying the
c.Opt.MaxSize value, which produced unexpected results for the user.