Do not compare more than MaxSize bytes
Ensure that reading a stream beyond MaxSize doesn't include bytes in the
comparison that are beyond the MaxSize amount.  Do this by using
LimitedReader to limit the amount of bytes read before the bytes.Equal()
comparison to no more than MaxSize.  After the comparison result is
determined, read from the original Reader to see if there were more
bytes beyond MaxSize, and return an error along with the equality value.