Ignore MaxSize when at least one reader is Limited
If one or both Readers are LimitedReader, ignore MaxSize, and return
false with no error, if the amount of read data from each reader is not
the same.

In order to determine if there is more data beyond either the limit
given by one LimitedReader, or the MaxSize if neither given Reader is a
LimitedReader, the CompareReader() func can read one beyond the given
limit.